# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ClinicCreate": {
        "properties": {
          "address_summary": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Summary"
          },
          "area": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "booking_policy": {
            "default": "provider_confirmation",
            "pattern": "^(instant|provider_confirmation)$",
            "title": "Booking Policy",
            "type": "string"
          },
          "city": {
            "maxLength": 120,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "name": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phone_public": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Public"
          }
        },
        "required": [
          "name",
          "city"
        ],
        "title": "ClinicCreate",
        "type": "object"
      },
      "ClinicOut": {
        "properties": {
          "address_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Summary"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "booking_policy": {
            "title": "Booking Policy",
            "type": "string"
          },
          "city": {
            "title": "City",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone_public": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Public"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "integer"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "city",
          "area",
          "address_summary",
          "phone_public",
          "booking_policy",
          "policy_version"
        ],
        "title": "ClinicOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderOut": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "specialty": {
            "title": "Specialty",
            "type": "string"
          },
          "verification_status": {
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "specialty",
          "bio",
          "verification_status"
        ],
        "title": "ProviderOut",
        "type": "object"
      },
...
      "PublicProviderOut": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "specialty": {
            "title": "Specialty",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "specialty",
          "bio"
        ],
        "title": "PublicProviderOut",
        "type": "object"
      },
      "SlotOut": {
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "start_at",
          "end_at",
          "status"
        ],
        "title": "SlotOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NaviMed API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/messages/webhooks/provider": {
      "post": {
        "operationId": "provider_webhook_v1_messages_webhooks_provider_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-provider-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Provider-Signature"
            }
          },
          {
            "in": "header",
            "name": "x-provider-event-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Provider-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Webhook",
        "tags": [
          "messaging"
        ]
      }
    },
...
    "/v1/owner/providers": {
      "get": {
        "operationId": "providers_v1_owner_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Providers",
        "tags": [
          "owner"
        ]
      }
    },
...
    "/v1/providers": {
      "get": {
        "operationId": "list_providers_v1_providers_get",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "specialty",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Specialty"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicProviderOut"
                  },
                  "title": "Response List Providers V1 Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/me/onboarding": {
      "post": {
        "operationId": "provider_onboarding_v1_providers_me_onboarding_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Provider Onboarding",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}": {
      "get": {
        "operationId": "get_provider_v1_providers__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProviderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}/appointments": {
      "get": {
        "operationId": "provider_appointments_v1_providers__provider_id__appointments_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Provider Appointments",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}/availability": {
      "get": {
        "operationId": "availability_v1_providers__provider_id__availability_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From At"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SlotOut"
                  },
                  "title": "Response Availability V1 Providers  Provider Id  Availability Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Availability",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}/clinics": {
      "get": {
        "operationId": "get_provider_clinics_v1_providers__provider_id__clinics_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClinicOut"
                  },
                  "title": "Response Get Provider Clinics V1 Providers  Provider Id  Clinics Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider Clinics",
        "tags": [
          "providers"
        ]
      },
      "post": {
        "operationId": "create_clinic_v1_providers__provider_id__clinics_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Clinic",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}/schedule": {
      "post": {
        "operationId": "add_schedule_v1_providers__provider_id__schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "weekday",
            "required": true,
            "schema": {
              "title": "Weekday",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start_minute",
            "required": true,
            "schema": {
              "title": "Start Minute",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "end_minute",
            "required": true,
            "schema": {
              "title": "End Minute",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "appointment_duration_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Appointment Duration Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Schedule V1 Providers  Provider Id  Schedule Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Schedule",
        "tags": [
          "providers"
        ]
      }
    },
    "/v1/providers/{provider_id}/schedule/generate-slots": {
      "post": {
        "operationId": "generate_slots_v1_providers__provider_id__schedule_generate_slots_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_date",
            "required": true,
            "schema": {
              "title": "Target Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Slots V1 Providers  Provider Id  Schedule Generate Slots Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Slots",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}